+++ benchmark/kernels/fused_moe_triton/benchmark_torch_compile_fused_moe.py
@@ -7,7 +7,7 @@ from torch.nn import functional as F
 from transformers import AutoConfig
 
 from sglang.benchmark.bench_utils import run_bench
-from sglang.srt.layers.moe.fused_moe_triton.fused_moe import (
+from sglang.srt.layers.moe.moe_runner.triton_utils.fused_moe import (
     fused_moe as fused_moe_triton,
 )
 from sglang.srt.model_executor.cuda_graph_runner import set_torch_compile_config
